set qos for the volume and validate the replica and total IOPS

+++ src/org/apache/cloudstack/storage/datastore/utils/AppInstanceInfo.java
@@ -88,6 +88,6 @@ public class AppInstanceInfo {
   public int replicaCount;
   public int size;
   @SerializedName("capacity_in_use")
-  public int capacityInUse;
+  public long capacityInUse;
  }
 }
